+++ CHANGELOG.md
@@ -2350,3 +2350,4 @@
 - [2026-06-26 09:33] feat(auth): add async auth stream provider
 - [2026-06-26 09:43] test(auth): add unit tests for auth business logic
 - [2026-06-26 09:45] refactor(cache): optimize data flow between domain and data layers
+- [2026-06-26 10:20] feat(network): implement shimmer loading skeleton in network